+++ esphome/components/web_server_idf/utils.cpp
@@ -1,5 +1,7 @@
 #ifdef USE_ESP_IDF
 #include <memory>
+#include <cstring>
+#include <cctype>
 #include "esphome/core/helpers.h"
 #include "esphome/core/log.h"
 #include "http_parser.h"
@@ -86,6 +88,36 @@ optional<std::string> query_key_value(const std::string &query_url, const std::s
 
   url_decode(val.get());
   return {val.get()};
+}
+
+// Helper function for case-insensitive string region comparison
+bool str_ncmp_ci(const char *s1, const char *s2, size_t n) {
+  for (size_t i = 0; i < n; i++) {
+    if (!char_equals_ci(s1[i], s2[i])) {
+      return false;
+    }
+  }
+  return true;
+}
+
+// Case-insensitive string search (like strstr but case-insensitive)
+const char *stristr(const char *haystack, const char *needle) {
+  if (!haystack) {
+    return nullptr;
+  }
+
+  size_t needle_len = strlen(needle);
+  if (needle_len == 0) {
+    return haystack;
+  }
+
+  for (const char *p = haystack; *p; p++) {
+    if (str_ncmp_ci(p, needle, needle_len)) {
+      return p;
+    }
+  }
+
+  return nullptr;
 }
 
 }  // namespace web_server_idf
